C UPDATES January 25
Mike passed line_filename to run_hexic through memory. line_filename contains the path to the spectral line parameter file. I need to change the code so that it doesn’t read it from the input parameter values in INPUT.txt. I’ve done the following changes:

Changed read_line.f90 to read the file that is passed to the routine through the header. I have removed the USE INPUT_PARAM because the filename now comes through the header.
Changed input_param.f90 to remove ATOM_PATH (which used to contain the path for the spectral line file).
Removed the line that specified the path to the file in INPUT.txt
Removed the code that read that same information from read_input.f90

2. Mike also added a synthesis_mode option to run HEXIC in synthesis mode. This is accomplished by calling hectic_invert (from hectic_invert_example) using the flag /SYNTHESIS_MODE). This passes along a value of 1 to the variable IN_MODE in run_hexic.f90. If /Synthesis_mode is not present, then a value of 0 is passed through IN_MODE to run_hexic.f90). Changes to the fortran code:

In run_hexic reflect that the mode is “inversion” when IN_MODE equals 0 and “synthesis” when IN_MODE equals 1. I changed the 2 IF statements that checked the value of MODE. Now they check the value of  IN_MODE.
Removed the variable MODE from input_params.f90
Removed the code that reads the MODE from read_input.f90
Removed the corresponding lines in INPUT.txt
The variable MODE does not seem to be used anywhere else in the code, but I should keep an eye on it, just in case. Seems to be running just fine now that I removed it.

3. The last change that Mike did in November was to pass the instrument filter profiles through the header, from the IDL routine. The parameter is optional. What happens if the user does not provide a set of filter profiles? Then Nfilts should be equal to Nwavelengths and the filter profiles should be calculated internally to be delta-Dirac functions (or not use them at all!)

I created an IDL routine to read the filters (read_filt.pro)
I call this routine from run_hexic_example and pass the filters to the fortran code through the header.
The way the intermediate C routine works, I have to transpose the filter array when it gets into run_hexic.f90. This should be changed by swapping the indexes in the C routine.
I commented out the call to read_filt.f90 in run_hexic and removed the routine from the distribution.
I removed the INPUT.txt file altogether, the input_param.f90 that defined the variables and the routine that reads them: read_input, since they are not necessary anymore.
I removed input_param, read_input and read_filt from CMakelists

LEFT TO DO:

-Check for consistency between size of filter array and information in LINE file
-Find out what’s going on with the name of the line file name when passed through the headers of the IDL to the C to the Fortran routines.
-What happens if the user passes no filter profiles to the code.